Fix texture transparency issue
I discovered that when a PNG is loaded, it was triggering the case at line no. 448, which wasn't making an RGBA texture in OpenGL.

I also included some minor changes so we are hard-specifying the bit depth of each channel.  I'd also thing it would be a good idea to only use one type of texture (only make RGBA textures) but that's for another later discussion.

Also, the case states for lines 446-452 and 454-460 are kind of moot now, since they all resolve to the same proc call I collapsed it into an if statement.

Closes #36